+++ src/BoostCoreCommandProvider.php
@@ -4,29 +4,17 @@ declare(strict_types=1);
 
 namespace SanderMuller\BoostCore;
 
-use Composer\Command\BaseCommand;
 use Composer\Plugin\Capability\CommandProvider;
-use SanderMuller\BoostCore\Commands\DoctorCommand;
-use SanderMuller\BoostCore\Commands\InitCommand;
-use SanderMuller\BoostCore\Commands\InstallCommand;
-use SanderMuller\BoostCore\Commands\NewCommand;
-use SanderMuller\BoostCore\Commands\ScanCommand;
-use SanderMuller\BoostCore\Commands\SyncCommand;
+use SanderMuller\BoostCore\Commands\CommandRegistry;
+use Symfony\Component\Console\Command\Command;
 
 final class BoostCoreCommandProvider implements CommandProvider
 {
     /**
-     * @return array<BaseCommand>
+     * @return array<Command>
      */
     public function getCommands(): array
     {
-        return [
-            new InitCommand(),
-            new InstallCommand(),
-            new ScanCommand(),
-            new SyncCommand(),
-            new DoctorCommand(),
-            new NewCommand(),
-        ];
+        return CommandRegistry::commands();
     }
 }